Split Resolver from its result
